- Fine tuning.
git-svn-id: svn+ssh://svn.morknet.de/usr/local/subversion/blz3/trunk@847 c0f852eb-cc16-0410-a65c-258990c0cc53

+++ bin/clobber.sh
@@ -2,11 +2,13 @@
 
 set BLZ3_HOME=$1
 
+test $BLZ3_HOME || exit 1
+
 cd $BLZ3_HOME
 
 echo "Cleaning files..."
 if ( -f src/Makefile ) then
-  make clean
+  make -C src clean
   rm -rf `find src -name Makefile` config.*
 endif
 test -f include_unix/blz3/autoconf.h && rm -f include_unix/blz3/autoconf.h
